Startet to use Controllers and Routes, fixed some things in the log in and sign up and organised the code more
