+++ server/routes/authRoutes.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const router = express.Router();
+const authController = require("../controllers/authController");
+
+
+
+router.post("/login", authController.login)
+router.post("/signup", authController.signup)
+
+module.exports = router;
